refactor: recipe로 스타일링 변경
