+++ packages/ui/src/imageUploader/components/imageInput/style.css.ts
@@ -1,4 +1,5 @@
 import { style, styleVariants } from "@vanilla-extract/css";
+import { recipe } from "@vanilla-extract/recipes";
 import { Color } from "../../../globalStyle/theme.css";
 
 export const basicInput = style({
@@ -30,13 +31,24 @@ export const label = styleVariants({
   none: [base],
 });
 
-export const wrapper = style({
-  position: "relative",
-  border: "1px solid",
-  borderColor: Color.border,
-  borderRadius: "8px",
-  padding: "8px",
-  display: "flex",
-  backgroundColor: Color.primary.default,
-  cursor: "pointer",
+export const labelStyle = recipe({
+  base: {
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
+    cursor: "pointer",
+    gap: "10px",
+    width: "100%",
+    color: Color.text.inverse,
+    fontSize: "10px",
+    border: "2px dashed",
+    borderRadius: "8px",
+  },
+  variants: {
+    border: {
+      active: { borderColor: Color.accent.active },
+      none: { borderColor: "transparent" },
+    },
+  },
 });
